requirements.txt
google-adk>=0.1.0
fastapi>=0.100.0
uvicorn>=0.22.0
google-cloud-firestore>=2.11.0
pydantic>=2.0.0
